requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.2
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.7b0
cfgv==3.3.0
click==8.0.1
distlib==0.3.2
Django==3.2.5
django-filter==2.4.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
filelock==3.0.12
identify==2.2.11
iniconfig==1.1.1
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathspec==0.8.1
phonenumbers==8.12.27
platformdirs==2.0.2
pluggy==0.13.1
pre-commit==2.13.0
py==1.10.0
pylint==2.9.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.0.4
virtualenv==20.6.0
wrapt==1.12.1
